requirements.txt
django
gunicorn
django-heroku
djangorestframework==3.11.0